refacor:removed duplicate code when implementing calendars
