+++ src/core/components/addCardLogo.ts
@@ -1,4 +1,4 @@
-export function addCardImg(container: HTMLDivElement, src: string) {
+export function addCardLogo(container: HTMLDivElement, src: string) {
     container.innerHTML = "";
     const img = document.createElement('img');
     img.classList.add('payment__card-img')
