+++ src/core/components/background-video.ts
@@ -1,5 +1,4 @@
 export function renderBackground(src: string): HTMLVideoElement {
-
     const video = document.createElement('video');
     video.classList.add('background-video')
     video.autoplay = true
