+++ src/core/components/header.ts
@@ -2,8 +2,8 @@ import Component from '../templates/components';
 import { PageIds } from '../../utils/types';
 import { langArr } from '../../utils/dataLang';
 import { LanguageArr } from '../../utils/types';
-import { ticketsData } from '../../utils/dataBuyTicket';
-import { ticketsDataRu } from '../../utils/dataBuyTicket';
+import { ticketsData } from '../../utils/dataAllTickets';
+import { ticketsDataRu } from '../../utils/dataAllTickets';
 import { createDropDownListTicket } from './drop-down-list';
 import { createDropDownListCondition } from './drop-down-list';
 import { conditionsData } from '../../utils/dataCondition';
@@ -14,7 +14,7 @@ import { createSchedulesTable } from './schedulesTable';
 import { schedulesSaturdayData } from '../../utils/schedulesData';
 import { schedulesSaturdayDataRu } from '../../utils/schedulesData';
 import PricesPage from '../../pages/prices-calendar/prices';
-import { createCalendarView } from './create-calendar';
+import { createCalendarView } from './create-calendar-price';
 import { createCalendarView as createCalendarSchedules } from './create-calendar-schedules';
 import { langArrBuyTicket } from '../../utils/dataLang';
 import { QuizResult } from '../../utils/types';
