+++ src/pages/prices-calendar/prices.ts
@@ -2,17 +2,17 @@ import Page from '../../core/templates/page';
 import './prices.css';
 import { renderBackground } from '../../core/components/background-video';
 import Video from '../../assets/video/video-Oceanografic_Valencia.mov';
-import { createCalendarView } from '../../core/components/create-calendar';
+import { createCalendarView } from '../../core/components/create-calendar-price';
 import { highlightCell } from '../../utils/highlightCells';
 import { createDropDownListTicket } from '../../core/components/drop-down-list';
 import { createDropDownListCondition } from '../../core/components/drop-down-list';
-import { ticketsData } from '../../utils/dataBuyTicket';
+import { ticketsData } from '../../utils/dataAllTickets';
 import { conditionsData } from '../../utils/dataCondition';
 import { freeEntriesData } from '../../utils/dataFreeEntries';
 import { langArr } from '../../utils/dataLang';
 import { langArrHeaderFooter } from '../../utils/dataLang';
 import { LanguageArr } from '../../utils/types';
-import { ticketsDataRu } from '../../utils/dataBuyTicket';
+import { ticketsDataRu } from '../../utils/dataAllTickets';
 import { conditionsDataRu } from '../../utils/dataCondition';
 import { freeEntriesDataRu } from '../../utils/dataFreeEntries';
 import { PageIds } from '../../utils/types';
